+++ app/delete-account/page.tsx
@@ -1,51 +1,16 @@
-'use client';
+import dynamic from 'next/dynamic';
+import React, { Suspense } from 'react';
 
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { LoadingArea } from '@/components/atoms';
 
-import { BackButton, HeaderBar } from '@/components/molecules';
-import Step1 from '@/features/setting/components/organisms/step-1';
-import Step2 from '@/features/setting/components/organisms/step-2';
-import Step3 from '@/features/setting/components/organisms/step-3';
+const StepComponent = dynamic(() => import('./index'), {
+  suspense: true,
+});
 
 export default function Page() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [step, setStep] = useState(1);
-
-  useEffect(() => {
-    const queryStep = searchParams.get('step');
-    if (queryStep) {
-      setStep(Number(queryStep));
-    }
-  }, [searchParams]);
-
-  const handleListItemClick = () => {
-    setStep(2);
-    router.push('?step=2');
-  };
-
-  const getStepComponent = (step: number) => {
-    switch (step) {
-      case 1:
-        return <Step1 onListItemClick={handleListItemClick} />;
-      case 2:
-        return <Step2 />;
-      case 3:
-        return <Step3 />;
-      default:
-        return null;
-    }
-  };
-
   return (
-    <div>
-      <HeaderBar>
-        <HeaderBar.LeftContent>
-          <BackButton />
-        </HeaderBar.LeftContent>
-      </HeaderBar>
-      {getStepComponent(step)}
-    </div>
+    <Suspense fallback={<LoadingArea />}>
+      <StepComponent />
+    </Suspense>
   );
 }
